BTN Matrix

+++ Core/Inc/debouncer.h
@@ -12,7 +12,7 @@
 
 #define DEBOUNCE_TIME	10
 
-void DEB_SetTime(uint8_t array_index, uint8_t time);
+void DEB_SetTime(uint8_t array_index, uint32_t time);
 void DEB_DebounceArray(uint8_t * data_debounced, uint8_t * data_new, uint8_t array_length);
 
 #endif /* INC_DEBOUNCER_H_ */
